Autorun for Orion files fixed

+++ src/Orion.cpp
@@ -369,7 +369,7 @@ bool OrionFileLoader::loadFile(const std::string& fileName, bool run)
         m_platform->reset();
         Cpu8080Compatible* cpu = dynamic_cast<Cpu8080Compatible*>(m_platform->getCpu());
         if (cpu) {
-            g_emulation->exec(cpu->getKDiv() * 3000000);
+            g_emulation->exec((uint64_t)cpu->getKDiv() * 3000000);
             cpu->setPC(begAddr);
         }
